+++ nope/lib/views/index.php
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <title>Admin</title>
     <link href="<?php echo path('lib/assets/css/font-awesome.min.css'); ?>" rel="stylesheet" />
+    <link href="<?php echo path('lib/assets/css/simplemde.min.css'); ?>" rel="stylesheet" />
     <link href="<?php echo path('lib/assets/css/app.min.css'); ?>" rel="stylesheet" />
     <script>
       window.NOPE_BASE_PATH = "<?php echo $request->getUri()->getBasePath() . '/' . $request->getUri()->getPath(); ?>";
@@ -23,6 +24,7 @@
     <script src="<?php echo path('lib/assets/js/lib/ng-file-upload.min.js'); ?>"></script>
     <script src="<?php echo path('lib/assets/js/lib/angular-drag-and-drop-lists.min.js'); ?>"></script>
     <script src="<?php echo path('lib/assets/js/lib/moment.min.js'); ?>"></script>
+    <script src="<?php echo path('lib/assets/js/lib/simplemde.min.js'); ?>"></script>
     <?php foreach($js as $file) { ?>
     <script src="<?php echo path($file); ?>"></script>
     <?php } ?>